datatypes => 32bits

+++ src/clazz/profile.py
@@ -8,11 +8,11 @@ def _labels(knn, split_idx):
     n_timepoints, k_neighbours = knn.shape
 
     y_true = np.concatenate((
-        np.zeros(split_idx, dtype=np.int64),
-        np.ones(n_timepoints - split_idx, dtype=np.int64),
+        np.zeros(split_idx, dtype=np.int32),
+        np.ones(n_timepoints - split_idx, dtype=np.int32),
     ))
 
-    knn_labels = np.zeros(shape=(k_neighbours, n_timepoints), dtype=np.int64)
+    knn_labels = np.zeros(shape=(k_neighbours, n_timepoints), dtype=np.int32)
 
     for i_neighbor in range(k_neighbours):
         neighbours = knn[:, i_neighbor]
@@ -20,18 +20,18 @@ def _labels(knn, split_idx):
 
     ones = np.sum(knn_labels, axis=0)
     zeros = k_neighbours - ones
-    y_pred = np.asarray(ones > zeros, dtype=np.int64)
+    y_pred = np.asarray(ones > zeros, dtype=np.int32)
 
     return y_true, y_pred
 
 
 @njit(fastmath=True, cache=True)
 def _calc_neigh_pos(knn):
-    ind, val = np.zeros(knn.shape[0], np.int64), np.zeros(knn.shape[0] * knn.shape[1],
-                                                          np.int64)
+    ind, val = np.zeros(knn.shape[0], np.int32), np.zeros(knn.shape[0] * knn.shape[1],
+                                                          np.int32)
 
-    counts = np.zeros(ind.shape[0], np.int64)
-    ptr = np.zeros(ind.shape[0], np.int64)
+    counts = np.zeros(ind.shape[0], np.int32)
+    ptr = np.zeros(ind.shape[0], np.int32)
 
     # count knn occurences
     for idx in range(knn.shape[0]):
@@ -58,13 +58,13 @@ def _init_labels(knn, offset):
     n_timepoints, k_neighbours = knn.shape
 
     y_true = np.concatenate((
-        np.zeros(offset, dtype=np.int64),
-        np.ones(n_timepoints - offset, dtype=np.int64),
+        np.zeros(offset, dtype=np.int32),
+        np.ones(n_timepoints - offset, dtype=np.int32),
     ))
 
     neigh_pos = _calc_neigh_pos(knn)
 
-    knn_labels = np.zeros(shape=(k_neighbours, n_timepoints), dtype=np.int64)
+    knn_labels = np.zeros(shape=(k_neighbours, n_timepoints), dtype=np.int32)
 
     for i_neighbor in range(k_neighbours):
         neighbours = knn[:, i_neighbor]
@@ -72,14 +72,14 @@ def _init_labels(knn, offset):
 
     ones = np.sum(knn_labels, axis=0)
     zeros = k_neighbours - ones
-    y_pred = np.asarray(ones > zeros, dtype=np.int64)
+    y_pred = np.asarray(ones > zeros, dtype=np.int32)
 
     return (zeros, ones), neigh_pos, y_true, y_pred
 
 
 @njit(fastmath=True, cache=True)
 def _init_conf_matrix(y_true, y_pred):
-    conf_matrix = np.zeros(shape=(2, 4), dtype=np.float64)
+    conf_matrix = np.zeros(shape=(2, 4), dtype=np.int32)
 
     # for label in (0, 1):
     tp = np.sum(np.logical_and(y_true == 0, y_pred == 0))
@@ -110,7 +110,6 @@ def _update_conf_matrix(old_true, old_pred, new_true, new_pred, conf_matrix):
     # Entries are symmetrical
     conf_matrix[1, :] = conf_matrix[0, ::-1]
 
-    # print (conf_matrix)
     return conf_matrix
 
 
@@ -207,7 +206,7 @@ def _update_labels(split_idx, excl_zone, neigh_pos, knn_counts, y_true, y_pred,
 @njit(fastmath=True, cache=True)
 def _fast_profile(knn, window_size, score, offset):
     n_timepoints = knn.shape[0]
-    profile = np.full(shape=n_timepoints, fill_value=-np.inf, dtype=np.float64)
+    profile = np.full(shape=n_timepoints, fill_value=-np.inf, dtype=np.float32)
 
     knn_counts, neigh_pos, y_true, y_pred = _init_labels(knn, offset)
     conf_matrix = _init_conf_matrix(y_true, y_pred)
